Added access control to maintenance related components.
