+++ src/App.js
@@ -14,8 +14,8 @@ import Inventory from "./components/Inventory/Inventory";
 import Header from "./components/Common/Header";
 import AddInventory from "./components/Inventory/AddInventory";
 import SingleEquipment from "./components/Inventory/SingleEquipment";
-import Maintenance from "./components/Maintenance";
-import AddJob from "./components/Inventory/AddJob";
+import Maintenance from "./components/Maintenance/Maintenance";
+import AddJob from "./components/Maintenance/AddJob";
 import Dashboard from "./components/Dashboard/Dashboard";
 import EditEquipment from "./components/Inventory/EditEquipment";
 import Departments from "./components/Inventory/Departments";
